requirements.txt
Flask==2.2.5
web3==6.11.3
requests==2.31.0
python-dotenv==1.0.1
SQLAlchemy==1.4.54
alembic==1.13.2
psycopg2-binary==2.9.9
gunicorn==20.1.0
pydantic==1.10.18
celery==5.3.6
pytest==7.4.4
pytest-asyncio==0.23.7
eth-typing==3.5.2
